Switch to absolute μs timestamps on Windows

+++ src/lib.rs
@@ -36,17 +36,20 @@ impl Ignore {
 
 /// A MIDI structure used internally by the class to store incoming
 /// messages. Each message represents one and only one MIDI message.
+/// The timestamp is represented as the elapsed microseconds since
+/// a point in time that is arbitrary, but does not change for the
+/// lifetime of a given MidiInputConnection.
 #[derive(Debug, Clone)]
 struct MidiMessage {
     bytes: Vec<u8>,
-    timestamp: f64
+    timestamp: u64
 }
 
 impl MidiMessage {
     pub fn new() -> MidiMessage {
         MidiMessage {
             bytes: vec![],
-            timestamp: 0.0
+            timestamp: 0
         }
     }
 }
